% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erum_abstract.R
\name{erum_abstract}
\alias{erum_abstract}
\title{eRum2016 conference abstract format.}
\usage{
erum_abstract(...)
}
\arguments{
\item{...}{Arguments to \code{rmarkdown::pdf_document}}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown:render]{render}}
}
\description{
Format for eRum2016 conference abstracts. Adapted from
\href{https://github.com/rstudio/rticles}{rticles package}.
}

